fix: reactoring to centralize data fetching
